projects/dublin-bus-pipeline/requirements.txt
pandas>=2.0.0
requests>=2.28.0
folium>=0.14.0
plotly>=5.15.0
jupyter>=1.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
sqlite-utils>=3.35
streamlit>=1.28.0
